% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{as_lik_matrix}
\alias{as_lik_matrix}
\title{Coerce one of \code{log_lik()}'s two matrix arguments, naming it if it is wrong}
\usage{
as_lik_matrix(value, d, arg, what)
}
\description{
Coerce one of \code{log_lik()}'s two matrix arguments, naming it if it is wrong
}
\keyword{internal}
