% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{autocov}
\alias{autocov}
\title{Autocovariance of a single chain at lags 0..n-1, via the FFT}
\usage{
autocov(y, n)
}
\description{
Autocovariance of a single chain at lags 0..n-1, via the FFT
}
\keyword{internal}
