% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{bind_sim_draws}
\alias{bind_sim_draws}
\title{Stack one-observation draws into an \verb{n x d} matrix}
\usage{
bind_sim_draws(draws, d = NULL)
}
\description{
Stack one-observation draws into an \verb{n x d} matrix
}
\keyword{internal}
