% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2st.R
\name{c2st_logistic_prob}
\alias{c2st_logistic_prob}
\title{Cross-validated logistic regression for \code{\link[=c2st]{c2st()}}}
\usage{
c2st_logistic_prob(x_train, y_train, x_test)
}
\arguments{
\item{x_train, y_train}{Training draws and their 0/1 labels.}

\item{x_test}{Draws to score.}
}
\value{
Predicted probability of class 1 for each row of \code{x_test}.
}
\description{
Cross-validated logistic regression for \code{\link[=c2st]{c2st()}}
}
\keyword{internal}
