% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_positive}
\alias{check_positive}
\title{Validate a strictly positive scalar}
\usage{
check_positive(x, arg)
}
\arguments{
\item{x}{The user's value.}

\item{arg}{Name of the argument.}
}
\value{
\code{x} as a double.
}
\description{
Validate a strictly positive scalar
}
\keyword{internal}
