% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_prob}
\alias{check_prob}
\title{Validate a probability argument}
\usage{
check_prob(p, arg, open = TRUE)
}
\arguments{
\item{p}{The user's value.}

\item{arg}{Name of the argument.}

\item{open}{Require \code{0 < p < 1} (the default). \code{FALSE} allows the endpoints.}
}
\value{
\code{p} as a double.
}
\description{
Validate a probability argument
}
\keyword{internal}
