% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{de_drop_net}
\alias{de_drop_net}
\title{The fit without its torch module, for R-level serialization}
\usage{
de_drop_net(fit)
}
\description{
The fit without its torch module, for R-level serialization
}
\keyword{internal}
