% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{fit_maf}
\alias{fit_maf}
\title{Train a MAF on standardized (theta, x)}
\usage{
fit_maf(
  theta,
  x,
  n_transforms = 5L,
  hidden = c(50L, 50L),
  max_epochs = 2000L,
  batch_size = 200L,
  lr = 5e-04,
  validation_fraction = 0.1,
  patience = 20L,
  n_restarts = 1L,
  clip_grad_norm = 5,
  embedding = NULL,
  seed = NULL,
  verbose = FALSE,
  device = "cpu"
)
}
\description{
Train a MAF on standardized (theta, x)
}
\keyword{internal}
