% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{hint_parallel}
\alias{hint_parallel}
\title{Tell the user once per session how to go parallel}
\usage{
hint_parallel(n)
}
\description{
Tell the user once per session how to go parallel
}
\keyword{internal}
