% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{mdn_chunk_lp}
\alias{mdn_chunk_lp}
\title{Log-density of a chunk of observations under each row's mixture}
\usage{
mdn_chunk_lp(mix, xs, p)
}
\value{
A \verb{(n_theta, n_chunk)} tensor.
}
\description{
Log-density of a chunk of observations under each row's mixture
}
\keyword{internal}
