% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdn.R
\name{mdn_log_prob_tensor}
\alias{mdn_log_prob_tensor}
\title{Per-row mixture log density (in standardized theta space), as a torch tensor.
\code{theta}: (b, p) tensor, \code{x}: (b, q) tensor.}
\usage{
mdn_log_prob_tensor(net, theta, x)
}
\description{
Per-row mixture log density (in standardized theta space), as a torch tensor.
\code{theta}: (b, p) tensor, \code{x}: (b, q) tensor.
}
\keyword{internal}
