% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{n_things}
\alias{n_things}
\title{Count a noun, pluralizing it}
\usage{
n_things(n, word)
}
\arguments{
\item{n}{The count.}

\item{word}{The singular noun.}
}
\description{
Count a noun, pluralizing it
}
\keyword{internal}
