% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{net_param}
\alias{net_param}
\title{Named parameter lookup that fails loudly rather than silently returning NULL}
\usage{
net_param(net, name)
}
\description{
Named parameter lookup that fails loudly rather than silently returning NULL
}
\keyword{internal}
