% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{nsbi_workers}
\alias{nsbi_workers}
\title{Number of workers the current future plan provides (1 = sequential)}
\usage{
nsbi_workers()
}
\description{
Number of workers the current future plan provides (1 = sequential)
}
\keyword{internal}
