% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_families.R
\name{param_label}
\alias{param_label}
\title{Name one parameter in an error message, by name where there is one}
\usage{
param_label(param_names, j)
}
\arguments{
\item{param_names}{Parameter names, or \code{NULL}.}

\item{j}{Parameter index.}
}
\description{
Name one parameter in an error message, by name where there is one
}
\keyword{internal}
