% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_backend}
\alias{progress_backend}
\title{Which reporter carries the updates: "progressr", "builtin", or "none"}
\usage{
progress_backend()
}
\description{
Which reporter carries the updates: "progressr", "builtin", or "none"
}
\keyword{internal}
