% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress_option}
\alias{progress_option}
\title{Resolve the \code{neuralsbi.progress} option to one of
"auto", "on", "off", "builtin", "progressr"}
\usage{
progress_option()
}
\description{
Resolve the \code{neuralsbi.progress} option to one of
"auto", "on", "off", "builtin", "progressr"
}
\keyword{internal}
