% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{sample_prior}
\alias{sample_prior}
\title{Draw samples from a prior}
\usage{
sample_prior(prior, n)
}
\arguments{
\item{prior}{An \code{nsbi_prior} object.}

\item{n}{Number of samples.}
}
\value{
An \verb{n x dim} matrix of parameter draws.
}
\description{
Draw samples from a prior
}
