% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{stan_addend}
\alias{stan_addend}
\title{\code{+ c} or \verb{- |c|}, so the generated arithmetic reads like arithmetic}
\usage{
stan_addend(x)
}
\description{
\code{+ c} or \verb{- |c|}, so the generated arithmetic reads like arithmetic
}
\keyword{internal}
