% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{stan_bounds}
\alias{stan_bounds}
\title{A \verb{<lower=, upper=>} clause, or nothing when neither side is finite}
\usage{
stan_bounds(lower, upper)
}
\arguments{
\item{lower, upper}{The bounds.}
}
\description{
A \verb{<lower=, upper=>} clause, or nothing when neither side is finite
}
\keyword{internal}
