% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{stan_mat_of}
\alias{stan_mat_of}
\title{Stan expression slicing a matrix block back out of \code{w}}
\usage{
stan_mat_of(blocks, key, wname = "w")
}
\description{
Stan expression slicing a matrix block back out of \code{w}
}
\keyword{internal}
