% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{within_support}
\alias{within_support}
\title{Test whether parameters lie within the prior support}
\usage{
within_support(prior, theta)
}
\arguments{
\item{prior}{An \code{nsbi_prior} object.}

\item{theta}{A matrix (or vector) of parameters.}
}
\value{
Logical vector, one entry per row of \code{theta}.
}
\description{
Test whether parameters lie within the prior support
}
