% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_theta_matrix}
\alias{as_theta_matrix}
\title{Coerce parameters/data to a numeric matrix with a known column count}
\usage{
as_theta_matrix(x, d = NULL)
}
\description{
Coerce parameters/data to a numeric matrix with a known column count
}
\keyword{internal}
