% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_estimator.R
\name{dmvnorm_chol}
\alias{dmvnorm_chol}
\title{Multivariate normal log density using a precomputed upper-Cholesky factor
(\code{R} such that \code{Sigma = t(R) \%*\% R}, i.e. \code{chol(Sigma)}).}
\usage{
dmvnorm_chol(x, mean, R, log = TRUE)
}
\description{
Multivariate normal log density using a precomputed upper-Cholesky factor
(\code{R} such that \code{Sigma = t(R) \%*\% R}, i.e. \code{chol(Sigma)}).
}
\keyword{internal}
