% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{neuralsbi-package}
\alias{neuralsbi}
\alias{neuralsbi-package}
\title{neuralsbi: Neural Simulation-Based Inference in R}
\description{
A native R implementation of neural simulation-based inference, focused on Neural Posterior Estimation (NPE). Given a prior over parameters and a simulator, 'neuralsbi' trains a conditional neural density estimator to approximate the Bayesian posterior, enabling amortized, likelihood-free inference. The design mirrors the workflow of the Python 'sbi' package but is implemented directly in R (using the 'torch' back end for neural density estimators) rather than wrapping Python. It targets applied researchers who want an approachable interface with sensible defaults and built-in posterior diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedroliman/neural.sbi}
  \item Report bugs at \url{https://github.com/pedroliman/neural.sbi/issues}
}

}
\author{
\strong{Maintainer}: neural.sbi authors \email{plima@rand.org}

Authors:
\itemize{
  \item neural.sbi authors \email{plima@rand.org}
}

}
\keyword{internal}
