% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{neuralsbi-package}
\alias{neuralsbi}
\alias{neuralsbi-package}
\title{neuralsbi: Neural Simulation-Based Inference}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A native R implementation of neural simulation-based inference, focused on Neural Posterior Estimation. Given a prior over parameters and a simulator, 'neuralsbi' trains a conditional neural density estimator to approximate the Bayesian posterior, enabling amortized, likelihood-free inference. Neural estimators run on the 'torch' back end. It targets applied researchers who want an approachable interface with sensible defaults and built-in posterior diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedroliman.github.io/neuralsbi/}
  \item \url{https://github.com/pedroliman/neuralsbi}
  \item Report bugs at \url{https://github.com/pedroliman/neuralsbi/issues}
}

}
\author{
\strong{Maintainer}: Pedro Nascimento de Lima \email{plima@rand.org} (\href{https://orcid.org/0000-0001-9057-198X}{ORCID})

Authors:
\itemize{
  \item Pedro Nascimento de Lima \email{plima@rand.org} (\href{https://orcid.org/0000-0001-9057-198X}{ORCID})
}

}
\keyword{internal}
