% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior_custom}
\alias{prior_custom}
\title{Build a prior from arbitrary sampling / density functions}
\usage{
prior_custom(sample_fn, log_prob_fn = NULL, dim, lower = NULL, upper = NULL)
}
\arguments{
\item{sample_fn}{Function \verb{function(n)} returning an \verb{n x dim} matrix.}

\item{log_prob_fn}{Function \verb{function(theta)} returning a length-\code{n} vector of
log densities. Optional; required only for methods/diagnostics that need it.}

\item{dim}{Number of parameters.}

\item{lower, upper}{Optional support bounds (numeric vectors) enabling
out-of-support rejection.}
}
\value{
An \code{nsbi_prior} object.
}
\description{
Build a prior from arbitrary sampling / density functions
}
