% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior_normal}
\alias{prior_normal}
\title{Independent normal prior}
\usage{
prior_normal(mean, sd = 1)
}
\arguments{
\item{mean}{Numeric vector of means (one per parameter).}

\item{sd}{Numeric scalar or vector of standard deviations.}
}
\value{
An \code{nsbi_prior} object.
}
\description{
Independent normal prior
}
\examples{
prior <- prior_normal(mean = c(0, 0), sd = 1)
}
