% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_ggplot2}
\alias{require_ggplot2}
\title{Check that ggplot2 (and, for \code{\link[=pairplot]{pairplot()}}, GGally) are available}
\usage{
require_ggplot2(ggally = FALSE)
}
\description{
Check that ggplot2 (and, for \code{\link[=pairplot]{pairplot()}}, GGally) are available
}
\keyword{internal}
