% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardizer_log_jac}
\alias{standardizer_log_jac}
\title{Log absolute Jacobian determinant of the \emph{inverse} standardization
(standardized -> original). Constant, so a scalar.}
\usage{
standardizer_log_jac(std)
}
\description{
Log absolute Jacobian determinant of the \emph{inverse} standardization
(standardized -> original). Constant, so a scalar.
}
\keyword{internal}
